requirements.txt
flask
python-decouple
requests
pymongo 
dnspython 